+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Last Changes
 
+- [#10](https://github.com/LaxarJS/laxar-tooling/issues/10): provide `artifactValidator` to apply JSON schemas
+  + NEW FEATURE: see ticket for details
 - [#8](https://github.com/LaxarJS/laxar-tooling/issues/8): generate only one theme per artifact listing
   + **BREAKING CHANGE:** see ticket for details
 - [#9](https://github.com/LaxarJS/laxar-tooling/issues/9): the `fileReader` and `jsonReader` utilities were removed